requirements.txt
fastmcp>=2.0
mcp>=1.0
openai>=1.40
gradio>=4.0
flask>=3.0
python-dotenv>=1.0
pydantic>=2.0
pytest>=8.0
